Berkan development

+++ routes/Controller/User.js
@@ -4,13 +4,23 @@ let sequelize = require('../Util/DatabaseConnection').getSequelize;
 let tableNames = require('../Util/DatabaseConnection').getTableNames;
 let mUserFunc = require('../Util/DatabaseConnection').getUserModel;
 
-let db = sequelize();
+let isAdmin = require('./RoleCheck').isAdmin;
+let isWaiter = require('./RoleCheck').isWaiter;
+let isBartender = require('./RoleCheck').isBartender;
+let isChef = require('./RoleCheck').isChef;
+let isMatre = require('./RoleCheck').isMatre;
+let errorMessage = require('./RoleCheck').errorMesage;
+
+
+    let db = sequelize();
 let dbNames = tableNames();
 let mUser = db.model(dbNames.user);
 let mRole = db.model(dbNames.role);
 let mTable = db.model(dbNames.table);
+let mUserRoles = db.model(dbNames.userRoles);
 
-module.exports = function(app) {
+module.exports = function(app,session) {
+
 
     /*
         app.use('/user', function (req, res, next) {
@@ -29,25 +39,32 @@ module.exports = function(app) {
         //response.render(path.resolve('../../public/Pages/index.html'));
     }),
         app.get('/api/user/deleteUser/:username', function (request, response ) {
-            //localhost:3000/api/getATableBelongToAUser/berkan
 
-            //response.writeHead(200, { 'Content-Type': 'application/json' });
             console.log("Delete USER");
-            var username = request.params.deleteUser;
 
-            deleteUser(username).then(user => {
-                response.statusCode = 200;
-                console.log(user);
-                response.write("Successful",()=>{
-                    response.end();
-                });
-            }).catch(error => {
-                response.statusCode(404);
-                console.log(error);
-                response.write(error,()=>{
-                    response.end();
-                });
-            })
+            if (session != undefined && isAdmin(session.roleId)) {
+
+                    var username = request.params.username;
+
+                    deleteUser(username).then(user => {
+                        response.statusCode = 200;
+                        console.log(user);
+                        response.write("Successful", () => {
+                            response.end();
+                        });
+                    }).catch(error => {
+                        response.statusCode = 404;
+                        console.log(error);
+                        response.write(error, () => {
+                            response.end();
+                        });
+                    })
+                } else {
+                    response.write(errorMessage(), () => {
+                        response.statusCode = 404;
+                        response.end();
+                    })
+                }
 
         }),
 
@@ -58,117 +75,151 @@ module.exports = function(app) {
             var password = request.params.password;
             */
             var data = request.body;
+            console.log(request.body);
 
             checkValidationOfUser(data.username,data.password).then(user => {
                 response.statusCode = 200;
                 console.log(user);
-
                 response.write("Successful",()=>{
+
+                    session.username=data.username;
+                    getAUserRole(data.username).then(role=>{
+                        var myRole = JSON.parse(role);
+                        console.log("myrole "+ myRole[0].roleId);
+                        session.roleId = myRole[0].roleId;
+                        console.log("Session: " + session.username + session.roleId);
+                    }).catch(error=>{
+                        console.log(error);
+                    })
                     response.end();
                 });
             }).catch(error => {
                 response.statusCode = 404;
                 console.log(error);
-
-                response.write(error,()=>{
+                response.write(error.toString(),()=>{
                     response.end();
                 });
             })
 
         }),
 
-        app.post('/api/user/addUser', function (request, response,next) {
+        app.post('/api/user/addUser', function (request, response) {
             console.log("Create A User");
             var data = request.body;
             console.log(data);
+                if (session != undefined && isAdmin(session.roleId)) {
+                    data.roleId = parseInt(data.roleId, 10);
+                    data.courseId = parseInt(data.courseId, 10);
 
-            data.roleId = parseInt(data.roleId,10);
-            data.courseId= parseInt(data.courseId,10);
 
+                    console.log(data);
+                    createAUser(data).then(user => {
+                        response.statusCode = 200;
+                        console.log(user);
 
-            console.log(data);
-            createAUser(data).then(user => {
-                response.statusCode = 200;
-                console.log(user);
+                        response.write("Successful", () => {
+                            response.end();
+                        });
+                    }).catch(error => {
+                        response.statusCode = 404;
+                        console.log(error);
 
-                response.write("Successful",()=>{
-                    response.end();
-                });
-            }).catch(error => {
-                response.statusCode = 404;
-                console.log(error);
-
-                response.write(error,()=>{
-                    response.end();
-                });
-            })
+                        response.write(error.toString(), () => {
+                            response.end();
+                        });
+                    })
+                } else {
+                    response.write(errorMessage(), () => {
+                        response.statusCode = 404;
+                        response.end();
+                    })
+                }
 
         }),
         app.post('/api/user/updateAUser', function (request, response) {
             console.log("Update A User");
-            response.setHeader('Content-Type', 'application/json' );
-            var data = request.body;
-            updateAUser(data).then(user => {
-                response.statusCode = 200;
-                console.log(user);
-                response.send(user);
-            }).catch(error => {
-                response.statusCode = 404;
-                console.log(error);
+                if (session != undefined && isAdmin(session.roleId)) {
+                    var data = request.body;
+                    updateAUser(data).then(user => {
+                        response.statusCode = 200;
+                        console.log(user);
+                        response.send(user.toString());
+                    }).catch(error => {
+                        response.statusCode = 404;
+                        console.log(error);
 
-                response.write(error,()=>{
-                    response.end();
-                });
-            })
+                        response.write(error.toString(), () => {
+                            response.end();
+                        });
+                    })
+                }else {
+                    response.write(errorMessage(), () => {
+                        response.statusCode = 404;
+                        response.end();
+                    })
+                }
 
         }),
 
 
 
         app.get('/api/user/getAllUsers/:username', function (request, response) {
+
             console.log("Get all Users");
+                if (session != undefined && isAdmin(session.roleId)) {
+                    getAllUsers().then(user => {
+                        response.statusCode = 200;
+                        console.log(user[0].get(0));
+                        response.write(user[0].get(0).toString(), () => {
+                            response.end();
+                        })
+                    }).catch(error => {
+                        response.statusCode = 404;
+                        console.log(error);
 
-            getAllUsers().then(user => {
-                response.statusCode = 200;
-                console.log(user[0].get(0));
-                response.write(user[0].get(0).toString(),()=>{
-                    response.end();
-                })
-            }).catch(error => {
-                response.statusCode = 404;
-                console.log(error);
-
-                response.write(error,()=>{
-                    response.end();
-                });
-            })
-
+                        response.write(error, () => {
+                            response.end();
+                        });
+                    })
+                }else {
+                    response.write(errorMessage(), () => {
+                        response.statusCode = 404;
+                        response.end();
+                    })
+                }
 
         }),
 
         app.get('/api/getATableBelongToAUser/:username', function (request, response) {
             console.log("getATableBelongToAUser");
-
-            var username = request.params;
-            console.log(username);
-            getATableBelongToAUser(username).then(data => {
-                response.statusCode = 200;
-                console.log(data);
-                response.write(data.toString(), () => {
+            if (session != undefined &&
+                (isAdmin(session.roleId) || isWaiter(session.roleId) || isChef(session.roleId)) ||isMatre(session.roleId))
+            {
+                var username = request.params;
+                console.log(username);
+                getATableBelongToAUser(username).then(data => {
+                    response.statusCode = 200;
+                    console.log(data);
+                    response.write(data.toString(), () => {
+                        response.end();
+                    })
+                })
+                    .catch(error => {
+                        response.statusCode = 404;
+                        console.log(error);
+                        response.write(error, () => {
+                            response.end();
+                        });
+                    })
+            }else {
+                response.write(errorMessage(), () => {
+                    response.statusCode = 404;
                     response.end();
                 })
-            })
-                .catch(error => {
-                    response.statusCode = 404;
-                    console.log(error);
-                    response.write(error, () => {
-                        response.end();
-                    });
-                })
-
+            }
         })
 
-
+}
 
     /*
       app.get('/api/:user/:addARole', function (request, response,next) {
@@ -181,23 +232,11 @@ module.exports = function(app) {
   })
 
 
-      app.get('/api/:user/:getAUserRole', function (request, response) {
-          console.log('getAUserRole');
-          getAUserRole('berkan').then(data => {
-              console.log(data);
-              response.write('2');
-          }).catch(error => {
-              console.log(error);
-              response.write('2');
-          })
-
-      })
-
 
 */
 
 
-}
+
 
 
 
@@ -310,17 +349,52 @@ function addARole(data){
     })
 }
 /*
-const getAUserRole = (username)=>{
-    return new Promise((resolve,reject)=>{
-        mUser.findByPk('username')
-            .then((user)=>{
-                resolve(user.getRoles())
-            }).catch(error=>{
+//Many to Many Example
+function getAUserRole(data){
+    console.log(data);
+    return new Promise((resolve, reject) => {
+        mUser.findAll({
+            where: {
+                username: data
+            },
+            include: [
+                {
+                    model:mRole,
+                    through: mUserRoles
+                }
+            ]
+        }).then(myData=>{
+            if (myData != null && myData != undefined){
+                resolve(JSON.stringify(myData));
+            }
+            else
+                reject("Cannot get the role Id!");
+        }).catch(error => {
             reject(error);
         })
+
     })
 }
 */
+function getAUserRole(data){
+    console.log(data);
+    return new Promise((resolve, reject) => {
+        mUserRoles.findAll({
+            where: {
+                userUsername: data
+            }
+        }).then(myData=>{
+            if (myData != null && myData != undefined){
+                resolve(JSON.stringify(myData));
+            }
+            else
+                reject("Cannot get the role Id!");
+        }).catch(error => {
+            reject(error);
+        })
+
+    })
+}
 
 function checkValidationOfUser(username, password){
     return new Promise((resolve, reject) => {
@@ -331,6 +405,8 @@ function checkValidationOfUser(username, password){
             }
         }).then(user=>{
             if (user != null && user != undefined ){
+                user = JSON.stringify(user);
+                //console.log(user);
                 resolve(user);
             }else{
                 reject("Username or Password is wrong!");
